requirements.txt
asttokens==2.4.1
blinker==1.9.0
boto3==1.35.56
botocore==1.35.56
cffi==1.17.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
dnspython==2.7.0
executing==2.1.0
Flask==3.1.0
Flask-Cors==5.0.0
flask-marshmallow==1.2.1
fonttools==4.54.1
greenlet==3.1.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.3
oracledb==2.5.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pymongo==4.10.1
PyMySQL==1.1.1
pyodbc==5.2.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyzmq==26.2.0
s3transfer==0.10.3
six==1.16.0
SQLAlchemy==2.0.36
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
